requirements.txt
# === facewood （仮想環境の名前） requirements ===
# PyTorch の CUDA 12.1 ビルドを使う（Windows可・CUDA同梱ホイール）
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.2.2
torchvision==0.17.2

# 顔認証バックボーン
facenet-pytorch==2.6.0

# 依存（facenet-pytorch の要件に合わせてピン止め）
numpy>=1.24,<2.0
pillow==10.2.0

# データ/前処理/設定
pandas>=2.0,<2.3
scikit-learn>=1.3,<1.6
h5py>=3.8,<4.0
PyYAML>=6.0,<7.0

# ユーティリティ
tqdm>=4.66
requests>=2.31
